Compute contributing units to table cells

This function computes (with respect to the raw input data) the indices of all contributing units to given cells identified by ids.

contributing_indices(prob, ids = NULL)

Arguments

prob
a sdcProblem object created with makeProblem()

ids
a character vector containing default ids (strIDs) that define table cells. Valid inputs can be extracted by using sdcProb2df() and looking at column strID. If this argument is NULL, the correspondig units are computed for all cells in the table.

Value

a named list where names correspond to the given ids` and the values to the row numbers within the raw input data.

Examples

# loading test data
data("microData1", package="sdcTable")

# specify hierarchies for `age` and `region`
dim_region <- hier_create(root = "Total", nodes = LETTERS[1:4])
dim_gender <- hier_create(root = "Total", nodes = c("male", "female"))
dl <- list(region = dim_region, gender = dim_gender)

# no variables holding counts, numeric values, weights or sampling
# weights are available in the input data

# using variable names is also possible
prob <- makeProblem(
  data = microData1,
  dimList = dl
)

df <- sdcProb2df(prob, dimCodes = "original")

# which units contribute to cell region = "A" and gender = "female"?

# compute the id ("0101")
df[region == "A" & gender == "female", strID]

#> [1] "0101"

# which indices contribute to the cell?
ids <- contributing_indices(prob = prob, ids = "0101")

# check
dataObj <- get.sdcProblem(prob, "dataObj")
rawData <- slot(dataObj, "rawData")
rawData[ids[["0101"]]]

#>    region gender freq
#> 1:      A female    1
#> 2:      A female    1

# compute contributing ids for each cell
contributing_indices(prob)

#> $`0000`
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100
#> 
#> $`0001`
#>  [1]  1  2 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 54 55 56 57
#> [26] 58 59 60 61 62 63 76 77 78 79 80 81 82 83 84 85 86 87 88 89
#> 
#> $`0002`
#>  [1]   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20  40
#> [20]  41  42  43  44  45  46  47  48  49  50  51  52  53  64  65  66  67  68  69
#> [39]  70  71  72  73  74  75  90  91  92  93  94  95  96  97  98  99 100
#> 
#> $`0100`
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20
#> 
#> $`0101`
#> [1] 1 2
#> 
#> $`0102`
#>  [1]  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20
#> 
#> $`0200`
#>  [1] 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45
#> [26] 46 47 48 49 50 51 52 53
#> 
#> $`0201`
#>  [1] 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39
#> 
#> $`0202`
#>  [1] 40 41 42 43 44 45 46 47 48 49 50 51 52 53
#> 
#> $`0300`
#>  [1] 54 55 56 57 58 59 60 61 62 63 64 65 66 67 68 69 70 71 72 73 74 75
#> 
#> $`0301`
#>  [1] 54 55 56 57 58 59 60 61 62 63
#> 
#> $`0302`
#>  [1] 64 65 66 67 68 69 70 71 72 73 74 75
#> 
#> $`0400`
#>  [1]  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94
#> [20]  95  96  97  98  99 100
#> 
#> $`0401`
#>  [1] 76 77 78 79 80 81 82 83 84 85 86 87 88 89
#> 
#> $`0402`
#>  [1]  90  91  92  93  94  95  96  97  98  99 100
#>