query information from objects

Function getInfo is used to query information from objects of class sdcProblem-class, problemInstance-class or safeObj-class

getInfo(object, type)

Arguments

object
a sdcProblem-class object, problemInstance-class object or safeObj-class object.

type
a character vector of length 1 specifying the information which should be returned.

if argument object is of class sdcProblem-class or problemInstance-class, valid choices are:

lb: slot 'lb' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

ub: slot 'ub' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

LPL: slot 'LPL' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

SPL: slot 'SPL' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

UPL: slot 'UPL' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

sdcStatus: slot 'sdcStatus' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

freq: slot 'freq' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

strID: slot 'strID' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

numVars: slot 'numVars' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

w: slot 'w' of input object if it is of class problemInstance-class or this slot within slot 'problemInstance' if object is of class sdcProblem-class

if argument object is of class safeObj-class, valid choices are:

finalData: slot 'finalData' of input object of class safeObj-class

nrNonDuplicatedCells: slot 'nrNonDuplicatedCells' of input object of class safeObj-class

nrPrimSupps: slot 'nrPrimSupps' of input object of class safeObj-class

nrSecondSupps: slot 'nrSecondSupps' of input object of class safeObj-class

nrPublishableCells: slot 'nrPublishableCells' of input object of class safeObj-class

suppMethod: slot 'suppMethod' of input object of class safeObj-class

Value

manipulated data dependend on arguments object and type

Examples

# load problem (as it was created in the example
# of \code{\link{makeProblem}})
sp <- searchpaths()
fn <- paste(sp[grep("sdcTable", sp)], "/data/problem.RData", sep="")
problem <- get(load(fn))

# problem is an object of class \code{\link{sdcProblem-class}}
print(class(problem))

#> [1] "sdcProblem"
#> attr(,"package")
#> [1] "sdcTable"

for (slot in c('lb','ub','LPL','SPL','UPL','sdcStatus',
  'freq', 'strID', 'numVars', 'w')) {
  cat('slot', slot,':\n')
  print(getInfo(problem, type=slot))
}

#> slot lb :
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> slot ub :
#>  [1] 200 110  90  40  36   5  66  28  38  44  24  20  50  22  28
#> slot LPL :
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> slot SPL :
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> slot UPL :
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> slot sdcStatus :
#>  [1] "s" "s" "s" "s" "s" "s" "s" "s" "s" "s" "s" "s" "s" "s" "s"
#> slot freq :
#>  [1] 100  55  45  20  18   2  33  14  19  22  12  10  25  11  14
#> slot strID :
#>  [1] "0000" "0001" "0002" "0100" "0101" "0102" "0200" "0201" "0202" "0300"
#> [11] "0301" "0302" "0400" "0401" "0402"
#> slot numVars :
#> $val
#>  [1] 1284  802  482  198  178   20  344  140  204  224  118  106  518  366  152
#> 
#> slot w :
#> NULL

# extracting information for objects of class \code{\link{safeObj-class}}
fn <- paste(sp[grep("sdcTable", sp)], "/data/protectedData.RData", sep="")
protectedData <- get(load(fn))
for (slot in c('finalData', 'nrNonDuplicatedCells', 'nrPrimSupps',
  'nrSecondSupps', 'nrPublishableCells', 'suppMethod')) {
  cat('slot', slot,':\n')
  print(getInfo(protectedData, type=slot))
}

#> slot finalData :
#>    region gender Freq  val sdcStatus
#> 1   Total  Total  100 1284         s
#> 2   Total   male   55  802         s
#> 3   Total female   45  482         s
#> 4       A  Total   20  198         s
#> 5       A   male   18  178         x
#> 6       A female    2   20         u
#> 7       B  Total   33  344         s
#> 8       B   male   14  140         s
#> 9       B female   19  204         s
#> 10      C  Total   22  224         s
#> 11      C   male   12  118         x
#> 12      C female   10  106         x
#> 13      D  Total   25  518         s
#> 14      D   male   11  366         s
#> 15      D female   14  152         s
#> slot nrNonDuplicatedCells :
#> [1] 15
#> slot nrPrimSupps :
#> [1] 1
#> slot nrSecondSupps :
#> [1] 3
#> slot nrPublishableCells :
#> [1] 11
#> slot suppMethod :
#> [1] "OPT"