create sdcProblem-class-objects

Function makeProblem is used to create sdcProblem-class-objects.

makeProblem(data, dimList, dimVarInd = NULL, freqVarInd = NULL,
  numVarInd = NULL, weightInd = NULL, sampWeightInd = NULL)

Arguments

data
a data frame featuring at least one column for each desired dimensional variable. Optionally the input data can feature variables that contain information on cell counts, weights that should be used during the cut and branch algorithm, additional numeric variables or variables that hold information on sampling weights.

dimList
a named list where the names refer to variable names in input data. Each list element can be one of:

tree: generated with hier_*() functions from the sdcHierarchies package

data.frame: a two column data.frame containing the complete level hierarchy of a dimensional variable using a top-to-bottom approach. The format of the data.frame is as follows:

first column: a character vector specifying levels with each vector element being a string only containing of '@'s from length 1 to n. If a vector element consists of i-chars, the corresponding code is of level i. The code '@' (one character) equals the grand total (level=1), the code `@@` (two characters) is of level 2 (directly below the overall total).

second column: a character vector specifying level codes

path: absolute or relative path to a .csv file that contains two columns seperated by semicolons (;) having the same structure as the "@;levelname"-structure described above

dimVarInd
numeric vector (or NULL) defining the column-indices of dimensional variables (defining the table) within argument data. If NULL, the names of argument dimList are used to calculate the indices of the dimensional variables within data internally.

freqVarInd
numeric vector (or NULL) defining the column-indices of a variable holding counts within argument data

numVarInd
numeric vector (or NULL) defining the column-indices of additional numeric variables available in argument data

weightInd
numeric vector of length 1 (or NULL) defining the column-index of a variable holding weights that should be used during as objective coefficients during the cut and branch algorithm to protect primary sensitive cells within argument data

sampWeightInd
numeric vector of length 1 (or NULL) defining the column-index of a variable holding sampling weights within argument data

Value

a sdcProblem-class-object

Examples

# loading micro data
data("microData1", package="sdcTable")
microData <- microData1; rm(microData1)

#> Warning: object 'microData1' not found

# having a look at the data structure
str(microData)

#> 'data.frame':	100 obs. of  3 variables:
#>  $ region: chr  "C" "C" "A" "A" ...
#>  $ gender: chr  "male" "male" "male" "male" ...
#>  $ val   : num  9 11 10 11 18 10 6 9 5 12 ...

# we can observe that we have a micro data set consisting
# of two spanning variables ('region' and 'gender') and one
# numeric variable ('val')

# specify structure of hierarchical variable 'region'
# levels 'A' to 'D' sum up to a Total
dim.region <- data.frame(
 levels=c('@','@@','@@','@@','@@'),
 codes=c('Total', 'A','B','C','D'),
 stringsAsFactors=FALSE)

# specify structure of hierarchical variable 'gender'
# using create_node() and add_nodes() (see ?manage_hierarchies)
dim.gender <- hier_create(root = "Total", nodes = c("male", "female"))
hier_display(dim.gender)

#> Total
#> ├─female
#> └─male

# create a named list with each element being a data-frame
# containing information on one dimensional variable and
# the names referring to variables in the input data
dimList <- list(region=dim.region, gender=dim.gender)

# third column containts a numeric variable
numVarInd <- 3

# no variables holding counts, numeric values, weights or sampling
# weights are available in the input data
freqVarInd <- weightInd <- sampWeightInd <- NULL

# creating an object of class \code{\link{sdcProblem-class}}
problem <- makeProblem(
 data=microData,
 dimList=dimList,
 freqVarInd=freqVarInd,
 numVarInd=numVarInd,
 weightInd=weightInd,
 sampWeightInd=sampWeightInd)

# what do we have?
print(class(problem))

#> [1] "sdcProblem"
#> attr(,"package")
#> [1] "sdcTable"

# have a look at the data
sdcProb2df(problem, addDups=TRUE,
  addNumVars=TRUE, dimCodes="original")

#>     strID freq sdcStatus  val region gender
#>  1:  0000  100         s 1284  Total  Total
#>  2:  0001   45         s  482  Total female
#>  3:  0002   55         s  802  Total   male
#>  4:  0100   20         s  198      A  Total
#>  5:  0101    2         s   20      A female
#>  6:  0102   18         s  178      A   male
#>  7:  0200   33         s  344      B  Total
#>  8:  0201   19         s  204      B female
#>  9:  0202   14         s  140      B   male
#> 10:  0300   22         s  224      C  Total
#> 11:  0301   10         s  106      C female
#> 12:  0302   12         s  118      C   male
#> 13:  0400   25         s  518      D  Total
#> 14:  0401   14         s  152      D female
#> 15:  0402   11         s  366      D   male