perform primary suppression in sdcProblem-class-objects

Function primarySuppression is used to identify and suppress primary sensitive table cells in sdcProblem-class objects. Argument type allows to select a rule that should be used to identify primary sensitive cells. At the moment it is possible to identify and suppress sensitive table cells using the frequency-rule, the nk-dominance rule and the p-percent rule.

primarySuppression(object, type, ...)

Arguments

object
a sdcProblem-class object

type
character vector of length 1 defining the primary suppression rule. Allowed types are:

freq: apply frequency rule with parameters maxN and allowZeros

nk: apply nk-dominance rule with parameters n, k and numVarInd

p: apply p-percent rule with parameters p and numVarInd

pq: apply pq-rule with parameters p and q

...
parameters used in the identification of primary sensitive cells. Parameters that can be modified|changed are:

maxN: numeric vector of length 1 used when applying the frequency rule. All cells having counts <= maxN are set as primary suppressed. The default value of maxN is 3.

allowZeros: logical vector of length 1 specifying if empty cells (count==0) should be considered sensitive when using the frequency rule. The default value of allowZeros is 'FALSE' so that empty cells are not considered primary sensitive by default.

p: numeric vector of length 1 specifying parameter p that is used when applying the p-percent rule with default value of 80.

pq: numeric vector of length 2 specifying parameters p and q that are used when applying the pq-rule with the default being c(25, 50).

n: numeric vector of length 1 specifying parameter n that is used when applying the nk-dominance rule. Parameter n is set to 2 by default.

k: numeric vector of length 1 specifying parameter k that is used when applying the nk-dominance rule. Parameter n is set to 85 by default.

numVarInd: numeric vector of length 1 specifying the index of the numerical variable that should be used to identify cells that are dominated by 2 (p-percent rule) or n (nk-dominance)-rule. If type is either 'nk', 'p' or 'pq', it is mandatory to specify numVarInd.

Value

a sdcProblem-class object

Note

the nk-dominance rule, the p-percent rule and the pq-rule can only be applied if micro data have been used as input data to function makeProblem.

Examples

# load micro data
sp <- searchpaths()
fn <- paste(sp[grep("sdcTable", sp)], "/data/microData1.RData", sep="")
microData <- get(load(fn))

# load problem (as it was created in the example in \code{\link{makeProblem}})
fn <- paste(sp[grep("sdcTable", sp)], "/data/problem.RData", sep="")
problem <- get(load(fn))

# we have a look at the frequency table by gender and region
xtabs(rep(1, nrow(microData)) ~ gender + region, data=microData)

#>         region
#> gender    A  B  C  D
#>   female  2 19 10 14
#>   male   18 14 12 11

# cell with region=='A' and gender=='female' has 2 units contributing to it
# this cell should be considered sensitive according the the freq-rule with 'maxN' equal to 2!
p1 <- primarySuppression(problem, type='freq', maxN=2)

# we can also apply a p-percent rule with parameter 'p' being 30 as below.
# This is only possible if we are dealing with micro data and we also have to specify the index of
# a numeric variable.
p2 <- primarySuppression(problem, type='p', p=30, numVarInd=1)

# looking at anonymization states we see, that one cell is primary suppressed (sdcStatus=='u')
# and the remaining cells are possible candidates for secondary suppression (sdcStatus=='s') given
# the frequency rule with parameter 'maxN=2'.
# Applying the p-percent rule with parameter 'p=30' resulted in two primary suppressions.
data.frame(p1.sdc=getInfo(p1, type='sdcStatus'), p2.sdc=getInfo(p2, type="sdcStatus"))

#>    p1.sdc p2.sdc
#> 1       s      s
#> 2       s      s
#> 3       s      s
#> 4       s      s
#> 5       s      s
#> 6       u      u
#> 7       s      s
#> 8       s      s
#> 9       s      s
#> 10      s      s
#> 11      s      s
#> 12      s      s
#> 13      s      s
#> 14      s      u
#> 15      s      s